Separated concepts of wallgrab and wallhug. ContactState.WALLGRAB is now ContactState.WALLADJACENT. wallHugging is a bool inside movement actuator. MovementController will set movementActuator to wall hug if moves into a wall while the contact state is adjacent. Will set movementActuator to stop hugging wall if no longer adjacent. Vertical speed is killed if hugging wall and has moved adjacent to the wall that frame.
Most things that key off WALLGRAB have been changed to key off movementActuator.GetIsHuggingWall(). The idea is that hugging a wall is now a deliberate action of the player: You have to be adjacent to a wall AND move into it. You don't automatically grab if you are flung into it, or jump while standing adjacent to it. Wall jumps still trigger when wall adjacent, though, in order to make them easy and reliable.