Increasing dead zone for analogue controllers to prevent inadvertent diagonal aiming. Added joypad controls to input manager. Removed down+jump causing a roll/dash. This was rarely useful, and frequently pressed inadvertently.
Spawners now have a limited number of spawns. They if they spawn characters, they keep a list of the corpuses of the spawned characters. If a spawner has spawned all of its objects, and every object with a corpus is dead, the spawner is said to be 'exhausted'. This will be used for determining if a player has cleared a stage or arena.